+++ package.json
@@ -38,15 +38,16 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
-    "@babel/preset-env": "^7.9.0",
-    "@babel/preset-react": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
     "@egy186/eslint-config": "^0.43.0",
-    "@material-ui/core": "^4.9.7",
+    "@material-ui/core": "^4.9.9",
     "@material-ui/icons": "^4.9.1",
     "axios": "^0.19.2",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "copyfiles": "^2.2.0",
-    "core-js": "^3.6.4",
+    "core-js": "^3.6.5",
     "eslint": "~6.8.0",
     "flat": "^5.0.0",
     "notistack": "^0.9.9",
@@ -63,7 +64,7 @@
     "rimraf": "^3.0.2",
     "terser-webpack-plugin": "^2.3.5",
     "use-persisted-state": "^0.3.0",
-    "webpack": "^4.42.0",
+    "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11"
   },
   "engines": {
